requirements.txt
pyspark>=3.5.1,<4.0
requests>=2.32.3,<3.0
google-cloud-storage>=2.16.0,<3.0
